requirements.txt
ipykernel
numpy
matplotlib
SQLAlchemy
panda
scipy